created weapon effects classes where gladiators may use it to damage opponents more

+++ src/Gladiators/BaseGladiator.cpp
@@ -2,20 +2,23 @@
 
 
 void BaseGladiator::increaseAbilities() {
-    SP = SP * (sp_multiplier * 0.01) * Level;
-    DEX = DEX * (dex_multiplier * 0.01) * Level;
+    SP = (int)(SP * (sp_multiplier * 0.01) * Level);
+    DEX = (int)(DEX * (dex_multiplier * 0.01) * Level);
     ++Level;
 }
 
 void BaseGladiator::healUp() {
     HP = StarterHealth;
-    HP = HP * (hp_multiplier * 0.01) * Level;
+    HP = HP * (int)(hp_multiplier * 0.01) * Level;
     StarterHealth = HP;
 }
 
-void BaseGladiator::decreaseHpBy(int decreaser) {
+void BaseGladiator::decreaseHpBy(int decreaser){
     HP = HP - decreaser;
     if(HP <= 0){
         isDead = true;
     }
 }
+
+
+
